Bug 813269 - Use double to represent time in Web Audio; r=bzbarsky

+++ content/media/webaudio/AudioContext.h
@@ -73,7 +73,7 @@ public:
   CreateGain();
 
   already_AddRefed<DelayNode>
-  CreateDelay(float aMaxDelayTime, ErrorResult& aRv);
+  CreateDelay(double aMaxDelayTime, ErrorResult& aRv);
 
   already_AddRefed<PannerNode>
   CreatePanner();
